Refactor : login 기능 리뷰 반영
