--- src/main/java/com/greedy/mokkoji/api/auth/dto/StudentInformationResponseDto.java
@@ -1,9 +1,0 @@
-package com.greedy.mokkoji.api.auth.dto;
-
-public record StudentInformationResponseDto(
-        String name,
-        String department,
-        String grade
-) {
-}
-
