+++ src/main/java/com/greedy/mokkoji/db/user/entity/User.java
@@ -3,9 +3,11 @@ package com.greedy.mokkoji.db.user.entity;
 import jakarta.persistence.*;
 import lombok.AccessLevel;
 import lombok.Builder;
+import lombok.Getter;
 import lombok.NoArgsConstructor;
 
 @Entity
+@Getter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @Table(name = "user")
 public class User {
