+++ src/main/java/com/greedy/mokkoji/enums/message/FailMessage.java
@@ -40,7 +40,8 @@ public enum FailMessage {
 
     //500
     INTERNAL_SERVER_ERROR(HttpStatus.INTERNAL_SERVER_ERROR, 50000, "서버 내부 오류가 발생했습니다."),
-    INTERNAL_SERVER_ERROR_SEJONG_AUTH(HttpStatus.INTERNAL_SERVER_ERROR, 50005, "학생 인증 서버 내부 오류가 발생했습니다.");
+    INTERNAL_SERVER_ERROR_SEJONG_AUTH(HttpStatus.INTERNAL_SERVER_ERROR, 50005, "학생 인증 서버 내부 오류가 발생했습니다."),
+    INTERNAL_TOKEN_INIT_ERROR(HttpStatus.INTERNAL_SERVER_ERROR, 50010, "SECRET KEY가 초기화되지 않았습니다.");
 
     private final HttpStatus httpStatus;
     private final int code;
